flask/requirements.txt
Flask
Gunicorn
Flask-Mail
requests
python-decouple
schedule
pandas
Flask-WTF
Flask-APScheduler
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary